+++ docs/src/examples/components/Popup/Types/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import ComponentExample from 'docs/src/components/ComponentDoc/ComponentExample'
+import ExampleSection from 'docs/src/components/ComponentDoc/ExampleSection'
+
+const Types = () => (
+  <ExampleSection title="Types">
+    <ComponentExample
+      title="Default"
+      description="A default popup."
+      examplePath="components/Popup/Types/PopupExample"
+    />
+  </ExampleSection>
+)
+
+export default Types