+++ src/components/Button/Button.tsx
@@ -1,7 +1,9 @@
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
+import whatInput from 'what-input'
+import * as _ from 'lodash'
 
-import { UIComponent, childrenExist, customPropTypes } from '../../lib'
+import { UIComponent, childrenExist, customPropTypes, createShorthandFactory } from '../../lib'
 import Icon from '../Icon'
 import { ButtonBehavior } from '../../lib/accessibility'
 import { Accessibility } from '../../lib/accessibility/interfaces'
@@ -12,6 +14,7 @@ import {
   ReactChildren,
   ComponentEventHandler,
 } from '../../../types/utils'
+import ButtonGroup from './ButtonGroup'
 
 export interface IButtonProps {
   as?: any
@@ -22,8 +25,11 @@ export interface IButtonProps {
   content?: React.ReactNode
   fluid?: boolean
   icon?: ItemShorthand
+  iconOnly?: boolean
   iconPosition?: 'before' | 'after'
   onClick?: ComponentEventHandler<IButtonProps>
+  onFocus?: ComponentEventHandler<IButtonProps>
+  text?: boolean
   type?: 'primary' | 'secondary'
   accessibility?: Accessibility
   styles?: IComponentPartStylesInput
@@ -33,15 +39,13 @@ export interface IButtonProps {
 /**
  * A button.
  * @accessibility
- * Default behavior: ButtonBehavior
- *  - adds role='button' if element type is other than 'button'
- *
- *
  * Other considerations:
  *  - for disabled buttons, add 'disabled' attribute so that the state is properly recognized by the screen reader
  *  - if button includes icon only, textual representation needs to be provided by using 'title', 'aria-label', or 'aria-labelledby' attributes
  */
 class Button extends UIComponent<Extendable<IButtonProps>, any> {
+  static create: Function
+
   public static displayName = 'Button'
 
   public static className = 'ui-button'
@@ -71,6 +75,9 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     /** Button can have an icon. */
     icon: customPropTypes.itemShorthand,
 
+    /** A button may indicate that it has only icon. */
+    iconOnly: PropTypes.bool,
+
     /** An icon button can format an Icon to appear before or after the button */
     iconPosition: PropTypes.oneOf(['before', 'after']),
 
@@ -80,6 +87,16 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
      * @param {object} data - All props.
      */
     onClick: PropTypes.func,
+
+    /**
+     * Called after user's focus.
+     * @param {SyntheticEvent} event - React's original SyntheticEvent.
+     * @param {object} data - All props.
+     */
+    onFocus: PropTypes.func,
+
+    /** A button can be formatted to show only text in order to indicate some less-pronounced actions. */
+    text: PropTypes.bool,
 
     /** A button can be formatted to show different levels of emphasis. */
     type: PropTypes.oneOf(['primary', 'secondary']),
@@ -104,9 +121,12 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     'disabled',
     'fluid',
     'icon',
+    'iconOnly',
     'iconPosition',
     'onClick',
+    'onFocus',
     'styles',
+    'text',
     'type',
     'variables',
   ]
@@ -114,6 +134,12 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
   public static defaultProps = {
     as: 'button',
     accessibility: ButtonBehavior as Accessibility,
+  }
+
+  static Group = ButtonGroup
+
+  public state = {
+    isFromKeyboard: false,
   }
 
   public renderComponent({
@@ -131,31 +157,24 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
         className={classes.root}
         disabled={disabled}
         onClick={this.handleClick}
+        onFocus={this.handleFocus}
         {...accessibility.attributes.root}
         {...rest}
       >
         {hasChildren && children}
-        {!hasChildren && iconPosition !== 'after' && this.renderIcon(variables)}
+        {!hasChildren && iconPosition !== 'after' && this.renderIcon()}
         {!hasChildren && content && <span className={classes.content}>{content}</span>}
-        {!hasChildren && iconPosition === 'after' && this.renderIcon(variables)}
+        {!hasChildren && iconPosition === 'after' && this.renderIcon()}
       </ElementType>
     )
   }
 
-  public renderIcon = variables => {
-    const { disabled, icon, iconPosition, content, type } = this.props
+  public renderIcon = () => {
+    const { icon, iconPosition, content } = this.props
 
     return Icon.create(icon, {
       defaultProps: {
         xSpacing: !content ? 'none' : iconPosition === 'after' ? 'before' : 'after',
-        variables: {
-          color:
-            type === 'primary'
-              ? variables.typePrimaryColor
-              : type === 'secondary'
-                ? variables.typeSecondaryColor
-                : variables.color,
-        },
       },
     })
   }
@@ -172,6 +191,16 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
       onClick(e, this.props)
     }
   }
+
+  private handleFocus = (e: React.SyntheticEvent) => {
+    const isFromKeyboard = whatInput.ask() === 'keyboard'
+
+    this.setState({ isFromKeyboard })
+
+    _.invoke(this.props, 'onFocus', e, this.props)
+  }
 }
+
+Button.create = createShorthandFactory(Button, content => ({ content }))
 
 export default Button