+++ test/specs/behaviors/Behavior-test.tsx
@@ -1,4 +1,14 @@
-import { ToggleButtonBehavior, IconBehavior } from '../../../src/lib/accessibility/'
+import {
+  ButtonBehavior,
+  InputBehavior,
+  ToggleButtonBehavior,
+  IconBehavior,
+  SelectableListItemBehavior,
+  BasicListBehavior,
+  BasicListItemBehavior,
+  SelectableListBehavior,
+  MenuBehavior,
+} from '../../../src/lib/accessibility/'
 import { TestHelper, TestMethod } from './testHelper'
 import definitions from './testDefinitions'
 const behaviorMenuItems = require('docs/src/behaviorMenu')
@@ -7,6 +17,13 @@ const testHelper = new TestHelper()
 testHelper.addTests(definitions)
 
 testHelper.addBehavior('ToggleButtonBehavior', ToggleButtonBehavior)
+testHelper.addBehavior('ButtonBehavior', ButtonBehavior)
 testHelper.addBehavior('IconBehavior', IconBehavior)
+testHelper.addBehavior('InputBehavior', InputBehavior)
+testHelper.addBehavior('SelectableListItemBehavior', SelectableListItemBehavior)
+testHelper.addBehavior('BasicListBehavior', BasicListBehavior)
+testHelper.addBehavior('BasicListItemBehavior', BasicListItemBehavior)
+testHelper.addBehavior('SelectableListBehavior', SelectableListBehavior)
+testHelper.addBehavior('MenuBehavior', MenuBehavior)
 
 testHelper.run(behaviorMenuItems)