+++ test/specs/behaviors/testDefinitions.ts
@@ -16,7 +16,6 @@ definitions.push({
   },
 })
 
-// Adds attribute 'aria-hidden=true' to icon.
 definitions.push({
   regexp: /Adds attribute '([\w\-\w \s*]+)=([a-z]+)' to [a-z]+/g,
   testMethod: (parameters: TestMethod) => {
@@ -24,6 +23,20 @@ definitions.push({
     const property = {}
     const expectedResult = parameters.behavior(property).attributes!.root[attributeToBeAdded]
     expect(expectedResult).toEqual(attributeExpectedValue)
+  },
+})
+
+definitions.push({
+  regexp: /Adds role='([a-z]+)'./g,
+  testMethod: (parameters: TestMethod) => {
+    const [_, roleToBeAdded] = [...parameters.props]
+    const property = {}
+    if (parameters.behavior instanceof Function) {
+      const expectedResult = parameters.behavior(property).attributes.root['role']
+    } else {
+      const expectedResult = parameters.behavior.attributes.root['role']
+    }
+    expect(expectedResult).toEqual(roleToBeAdded)
   },
 })